requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
async-to-sync==0.2.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.1
django-celery-results==2.5.1
django-filter==23.5
django-redis==5.4.0
djangorestframework==3.14.0
frozenlist==1.4.1
idna==3.6
kombu==5.3.5
magic-filter==1.0.12
multidict==6.0.4
outcome==1.3.0.post0
pdf2image==1.17.0
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
pytz==2023.4
redis==5.0.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
